Initial commit "Tools"
The tools folder is intended for anything extra we find or create that
will be useful for our project. for this commit, I've added a highly
specialized tally counter. Pass in the two text files and it will count up
and tally up the tags, combine them into one dictionary, and then output
it as a text file.